Update with scatter op

+++ keras_core/metrics/iou_metrics.py
@@ -127,7 +127,7 @@ class _IoUBase(Metric):
             y_pred,
             self.num_classes,
             weights=sample_weight,
-            dtype=self.dtype,
+            dtype="float32", # Important for weight assignment
         )
 
         return self.total_cm.assign(self.total_cm + current_cm)
